Js code modularized, slider ready.
