+++ src/experience.html
@@ -15,7 +15,6 @@
         href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,400,0,0" />
     <title>Home</title>
     <link rel="stylesheet" href="scss/styles.scss">
-    <script type="module" src="js/main.js"></script>
     <title>Experience</title>
 </head>
 
@@ -67,6 +66,7 @@
     </header>
 
     <!-- MAIN -->
+    <!-- main-section position relative -->
     <main class="container-fluid m-0 p-0 main-section">
         <div class="row m-0 p-0">
             <div class="col-12 col-lg-3 p-0 m-0">
@@ -209,6 +209,7 @@
         </div>
         <p>© MAB - All Rights Reserved.</p>
     </footer>
+    <script type="module" src="js/experience.js"></script>
 </body>
 
 </html>
